+++ changelog.md
@@ -13,6 +13,8 @@
 
   * Derive `Generic` and `Generic1` for `Data.Tree.Tree`.
 
+  * Add `intersperse` for sequences.
+
 ## 0.5.6.2  *Dec 2014*
 
   * Bundled with GHC 7.10.1.